pyproject.toml
[project]
name = "mp4towhisper"
version = "0.1.0"
description = "MP4/MP3 轉 SRT 字幕工具，使用 OpenAI Whisper 進行語音辨識"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "openai-whisper>=20250625",
    "opencc-python-reimplemented>=0.1.7",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu121"]
